Cache certain GitHub requests (#381)
* Cache certain GitHub requests

To avoid GH rate limit issues with big big clients we can cache some of our more used requests
that are unlikely to change in a short period of time.

This work was done initially on PR #379 by @trent-codecov.
The present changes are an attempt to integrate the changes better into our existing code.
In oarticular we wanted to use `OurOwnCache`, but share it for torngit instances.

I thought it was better to have a shared `TorngitCache` among all the torngit instances from a process.
This is not a problem, as the configuration of the cache doesn't change within a process, and so
`TorngitCache` was born. It is configured (for real) by the first instance, but every instance calls
the function (that then does nothing).

The more troublesome portion was that the original PR had logging for cache hits.
`OurOwnCache` is crystal clear fully transparent, and it's very versatile as well,
so it was a bit tricky to get info from it while making sure we don't log dangerous things.

The solution I landed on is to let each cached function specfity if the cache should log hits, and in this case
what arguments from the function should be called. This is as good a logging as we can get, I believe.

Notice from the tests `shared/tests/unit/helpers/test_cache.py::TestCache::test_simple_caching_fake_backend_with_params`
that there is a difference if you use args or kwargs. And the cache understands this difference as 2 separate func calls
(so it becomes 2 different keys). There's a difference in logging as well.

In any case this work was a bit rushed, and that was the best balance I could get to.
If you think this is a bit overkill, well I thought so to, but it's not in a bad shape.
And it's very easy to use.

No Tickets associated

* Make `init_torngit_cache` a method of the singleton.

Moving logic from `init_torngit_cache` to the `TorngitCache` so it's a method of the singleton.
The new method is `initialize`.