Extend `ReportDetails` to save data into GCS (#1180)
* Extend `ReportDetails` to save data into GCS

We have decided to move some data to GCS. Read more [here](https://www.notion.so/sentry/Moving-data-to-GCS-7fd3d7130a4d43bf987e55fe3090fc26?pvs=4). `ReportDetails.files_array` is the first column that we want to move.

The first step was to extend the model in codecov-api ([codecov-api_PR1543](https://github.com/codecov/codecov-api/pull/1543)) and do the migration to prepare the database for this change. Now we can extend the model here in the worker and actually start saving data into GCS.

Notice that the actual saving is feature-flagged and configurable under "setup" > "save_report_data_in_storage" > "report_details_files_array" (setting it to True). The idea is to have a new key for all columns we move, so that we can roll them out one by one and control exactly what data goes where. Furthermore, by setting "setup" > "save_report_data_in_storage" > "only_codecov" to True only repos that belong to codecov will save whatever columns we set to save in GCS in GCS. Other customers'data would still go to the DB.

You'll also notice that, differently from codecov-api (which uses Django's `@cached_property`) this one uses a mix of `@property` and `@lru_cache`. This is because `functools.cached_property` does accept writes, but I found the [docs](https://docs.python.org/3/library/functools.html#functools.cached_property) confusing and with a bunch of warnings. Then right at the end they suggest:

> If a mutable mapping is not available or if space-efficient key sharing is desired, an effect similar to cached_property() can also be achieved by stacking property() on top of lru_cache().

I thought that was a better option, since we can use `property()` and specify what is the getter and setter function. Plus we only really need to cache the getter function. In my opinion this made the code clearer and easier to understand, if a bit more verbose. And we have to remember to clear the getter cache when seting a new value, but it's not so bad. Plus, as the tests show, the cache does work. Nice.

* Include list of allowed repos

Similar to what was done for the `report_builder` process (see 52e7061617b00bd9a3473de8c0000854804afa21) we're adding
`"setup" > "save_report_data_in_storage" > "repo_ids"` as a list of repos for which saving files_array in storage is
turned on.

This will allow for a selective roll-out.