Refactor label analysis (#1195)
The present changes are a (big) refactor on the label analysis task.
The idea here is to group together and better encapsulate the major data-fetching processes that
we have with label analysis, namely getting and parsing the diff to extract relevant lines, and getting
the base report.

The operations above can be done without the labels from the request.
This is relevant because since [this change](https://github.com/codecov/codecov-api/pull/1585) we
can have the task run 2/3 of it without the labels, and have the labels patched later.
THESE CHANGES ASSUME WE DON'T DO THAT, but it's a first step to have that functionality.

So take this as a first step that will organize the code better before making further changes.

Another change I hope to make soon is to split `_get_base_report` and `_get_lines_relevant_to_diff`
to separate threads (using `threading` lib) and have them run concurrently. I think this makes sense
because they are both quite IO-bound.