Add network section to readable parsed raw report
This commit adds a loop that goes through the list of file names that make up
the network and adds them to the buffer containing the readable contents of the
raw report in the `content` method of the `ParsedRawReport` class.

This commit also resets the `toc` file pointer in the `LegacyParsedRawReport`
class.

Fixes: https://github.com/codecov/platform-team/issues/81

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>