PR comment mention diff between PR base and head
This commit  adds a line to the PR comment:

"Report is commits behind head on "

Where behind by is the difference between the PR base
and the base's branch's head in number of commits.
This commit adds a get_behind_by method to the ComparisonProxy
object. This method is called in the create_message method
in the MessageMixin class which assembles the PR comment.

The line will only be added to the PR comment if the
behind_by returned by the ComparisonProxy is not None
and is greater than 0.

get_behind_by first gets information about the base branch
via the get_branches method in the repository service, then
 compares the base branch's commit to the base

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>