Create `ArchiveField` (#15)
These changes are similar to https://github.com/codecov/codecov-api/pull/1609

The idea is that we want a wrapper to handle the readong/writing to GCS transparently to make it easier to do that in more models.
For that purpose we use a Descriptor, the `ArchiveField`.

Notice that it only works if the model specifies `__tablename__` but I think they all do. And it's easy to define if they don't.
Other changes are to adjust test cases.