Introduce `LabelAnalysisProcessingError` (#8)
Currently it's very hard for ATS users to self-support if the basic setup is complete.
Even if the minimal set of data to run ATS is present, it might happen that we miss some information and can't properly make a prediction. If we can't make a prediction we fallback to running all tests to avoid desrupting the user's CI run. Usually that's because we are missing static analysis info.

However it's hard for support to check this, because we need to check the task's logs. And it's impossible for the user to check that too. The effect of this is that the user will see all labels as absent, but won't have info to fix it.

By having `LabelAnalysisProcessingError` we can pass more information for the user, so that they know what data to send and can self-support. Hopefully this will make ATS better to use.