fix: change cache of `ArchiveField` (#55)
These changes are similar to https://github.com/codecov/codecov-api/pull/72
Same reasoning applies.

The test changed happens because now we update the cache on the write. Because of that we
are not doing the encode/decode/rehydrate operations. So the data you put in is the data you get.
On one hand we can do such operations to guarantee consistency. On the other this is no different than
what we used before `ArchiveField` AND such operations might be expensive.