fix: use committer date in github (#32)
context: codecov/codecov-api#97

Recently we had a situation in which a commit with a future date was not handled
properly by the system.

After some investigation the issue (obviously) was around the date. Turns out
we only update the HEAD commit of a PR if the candidate's timestamp is newer
than the HEAD's timestamp. For a commit in the future that poses a problem
because it will never be replaced with another HEAD (unless the future becomes present)

Git declares 2 different dates for commits, AuthorDate and CommitDate [read more](https://stackoverflow.com/questions/11856983/why-is-git-authordate-different-from-commitdate/11857467#11857467).
With GitHub we were usign the AuthorDate (which can be changed by `--date` arg).

My understanding is that the `CommitDate` is not user-configured (usually, but apparently there's an env var for that),
HOWEVER it get's updated by rebases, amends, force-pushes, etc.

We _only_ set the commit date when we first receive an upload for it, and never again.
In most cases this will be equal to the author date.
In the particular case of a future AuthorDate it will be the present-ish date when the commit was made.
I do believe that even with the changes to CommitDate using that instead of AuthorDate will
give us more confidence of a linear branch history.

In fact, we use the CommitDate with [gitlab](https://github.com/codecov/shared/blob/dd1e0c1aefb99c2286708d10f2e4f6c338bacf9e/shared/torngit/gitlab.py#L747)

I do believe that using this date will fix the PR Head update issue that we had,
without introducing new bugs to that logic,
and would have handled the force-push correctly and then not merged that future commit.

fix codecov/codecov-api#97