ref: drop ownerprofile terms agreement columns (#154)
Remove two columns from ownerprofile table that is no longer used, current data in those columns are migrated to the new columns in the user table, though the data is not meaningful at the moment anyways until we turn on the agreements feature on the frontend. Also fixes a bug where the owner has null userid due to impersonated login.