fix: Timeout Github API student check (#277) (#84)
Add a short timeout to the GitHub education API we use to check if a user is student. This GH call is known to be flaky, causing timeouts when this is called during the login step. This change gives it 3s to respond before default to "not a student" like how it is being handled currently when there's other errors coming from this GH call.