Create `Report.header` and populate it with labels_index (#94)
Changes the `CoverageDatapoint` to have `label_ids` instead of `labels`.

Notice that for many functions that expected labels I simply switched.
This is because we can adjust the reports when loading them up.

Also, many of the functions are related to upload processing, and we can fix the
reports when processing (on the worker side). So it's ok to expect that no report
with labels will get to those functions.

I was considering if I should make the `ReportLine` fix the labels when being generated, but decided against that.
The reason being that the label_idx would not be exclusively accessed. So 2 different workers might "fix" the same
report, but for whatever reason return a different index for it. This would cause the report to be inconsistem.

Of course this can be solved by using a proper hash for the index instead of numbers, but numbers save a lot more space. Instead we can trust the code that does "dangerous" operations (e.g. merge reports) to guarantee that both reports are in a good representation of labels before doing what needs to be done.

* chore: add label_index to `Report`

Part of the move with moving labels to an index table requires that the table exists.
It makes sense that it should be a part of the Report, especially during processing.

Introduces a report header, which is saved in chunks.txt.
Currently it will only hold the labels-index, but it's JSON,
so it can save more info.

* chore: Remove duplicates when merging datapoints

I noticed during testing that it's possible to have duplicated datapoints
when merging them. This happens if you have multiple report files uploaded
in one go.

That is useless and just takes up space. So by removing these duplicates we
save space and increase performance, even if by a negligible amount.

Not using a set to remove duplicates because the datapoint isn't hasheable.
But it's string representation works for the purposes of spotting duplicates

```
>>> l1 = (0, 1, None, [])
>>> str(l1)
'(0, 1, None, [])'
>>> S = set()
>>> S.add(l1)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: unhashable type: 'list'
```

* be smarter when scanning for the END_OF_HEADER

As pointed out by @scott-codecov, we were possibly scanning the
entire chunks file for END_OF_HEADER. But it's always at the 2nd
line of the file - if present.

We can be smarter about looking for it.