feat: Add component filtering for coverageFile type in GQL (#292)
In GraphQL API when we query the coverageFile type add an option in the input to filter by list of component names, This will work in conjunction with filtering by list of flags. If no coverage for the given components filters will return 0s for totals similar to if no coverage for the given flags. When filtering by both flags and components the result will be an intersection between flags and components.