feat: call CleanLabelsIndexTask (#210)
If the processed upload includes a flag that is carryforward
AND "labels" mode, we enqueue the CleanLabelsIndexTask.

Depends on https://github.com/codecov/worker/pull/205

I tried to be conservative, but it's possible that - despite
using the flag with labels - no labels are uploaded. It's
also possible that there's nothing to do. But these are small
details.

ps.: To see that indeed we have an "arguments" and "flag" in
"processing_results" I looked at logs for the message
"Scheduling notify task" :wink:

Thanks @matt-codecov for the suggestions around 
refactoring the trigger logic for CleanLabelsIndex to
make it easier to read and maintain.

closes codecov/engineering-team#768