feat: some per-customer ats metrics (#226)
context: codecov/engineering-team#526

Adds some basic per-customer ats metrics.
Actually because the `MetricsContext` usage is disabled this
just sets up the metrics once that's done.

Makes use of `MetricsContext` to annotate metrics per customer
(repo, commit, owner).

* replace `log_simple_metric` with `attempt` func

Replaces the `log_simple_metric` by the non-blocking, background
version `attempt_log_simple_metric`. I suppose it's more important
to continue the task than to wait for the metric, even if the diff
is almost irrelevant.

> **⚠️ Note on async operation**
>
> It should be safe to use `attempty_log_simple_metric`,
> given that all tasks run in an async function. Tests were
> not affected because all metrics are mocked. But the actual
> writing metrics behavior is untested in these changes.

* be explicit about using class attrs

Because the task classes persist to the process's duration
it's important to clean the state between task calls and
be explicit about it.