Bundle analysis: auto open and close DB session (#129)
Add a context manager for opening and closing DB sessions around queries in AssetReport and BundleReport. This way when the session in BundleAnalysisReport is closed, it won't make its children reports' session closed. This is needed in API when BundleAnalysisReport is resolved and cleaned up (ie DB session closed), but subsequent resolves to AssetReport and BundleReport will fail due to the shared DB session. This makes sure each *Report can manage its own DB session.