feat: Get GH installation token from multi apps (#131)
These changes add the ability to pass an `app_id` and a new location
for the PEM file for github installations.
This means that we can try to get tokens for installations other than
the default one (necessary to use multi apps).

The `app_id` is configured in the DB (in the installation model),
the pem is a little different in that it needs some manual config
(at least for now). Since we already mount the PEM for the default
app in the pods I figured we could just do the same initially to test
the feature and whatnot.

Realistically the value in the future will either be a path to a file
or an URL to some secret manager. It'll be easy to expand the current
implementation to support these other schemas in the future.

context: codecov/engineering-team#1146