feat: ramp up process for ghapp installations (#356)
* feat: ramp up process for ghapp installations

These changes introduce a change in weight for installations recently updated.
Any change in a ghapp `updated_at` field will cause it to enter a "ramp up" phase
during which other suitable installations will be prefered for usage.

The idea is that changes to an installation can introduce config errors and cause
the ghapp to be unusable in it's current state. To avoid breaking the app we use
the ramp up process.

The rampup is based on `random.choices` and giving weights to installs based on how
long ago they were last updated. The rap up function is exponential and the whole
process stabilizes after 10 days (at which point all installations have the same
probability of being selected)

closes: codecov/engineering-team#1483

* address review comments + fix tests