fix: bundle deletion deadlock (#169)
Attempt to fix a deadlock error when parsing bundle analysis reports. Remove the cascade delete of Asset from Chunks so that it doesn't become circular when Asset is cascade deleting Chunks. Similarly with Chunks-Modules. Flush after every delete operation to DB so that next query can be up to date.