fix: clean up pre-process task (#377)
This was originally motivated by [this issue](https://l.codecov.dev/MHMHKY).

Basically it was possible to call `services.repository.possibly_update_commit_from_provider_info`
passing an empty `repository_service`. That errors if we do have to update the commit info.
The case in which this happens is when the repo is not properly configured (no bot).

From that I removed some exceptions in `PreProcessUpload.get_repo_service` because it was impossible
to get those exception from the code inside the `try` block. I believe those were improperly copied from
some place else (like the places where I removed unused imports). All this work was originally done in the
`Upload` task, so possibly from there (the updating commits and initing reports is skipped in `Upload` if `PreprocessUpload` runs correctly).

Now breaking the task if we can't get a repository service. This is because if there are flags to carryforward
we need the commit information (to get the commit's parents and carryforward from there).
So allowing the task to continue could mean we have improperly carried forward reports.