fix: always select fallback if available (#422)
* fix: always select fallback if available

Because `random.choices` makes selections with replacement and we were making 3 selections only
there was a chance that an app would vanish from the choices (from not being selected).
This is bad because it hides this app as a possible fallbakc option.

Particularly if the choice is marked currently as "rate limited" we would fail with exception
instead of falling back in the other app. We want the fallback.

These changes go around this "choice with selection" by removing the selected app from the
population after each step. It's maybe slower (in practice the population will be small enough
that this doesn't really matter), but makes sure that we always select all available apps.
And still respectes the weighted selection.

* refactor: address review comments

Add docstrings to `_get_apps_from_weighted_selection` and refactor it
to make the code easier to understand. And reduce code complexity.