feat: extend ConnitNotification (#225)
* feat: extend ConnitNotification

These changes extend the `CommitNotification` model to include an optional
field: the github app that emitted a check / comment.

ticket: codecov/engineering-team#1737

* improve README with instructions on how to run migrations

* chore: include instance ID in GithubInstallationInfo

Up to this point we didn't pass the GithubAppInstallation ID to
the Torngit adapter. This was in part because the info was not needed,
and in part because `Owner.integration_id` doesn't have an ID.

It would be useful to have this info now, though. As it will allow
us to pin down an app to a specific app installation more easily.