Do not expose `BundleAnalysisReport.db_session` (#309)
* Properly `cleanup` all the `BundleAnalysisReport`s

* Avoid an `N+1` query

`bundle_reports` internally runs a query, so doing it in a loop is a bad idea.
As the second report is immutable, we can query it once ahead of time, turning this into just a single query.

* Do not expose `BundleAnalysisReport.db_session`

Instead of keeping a long-running `db_session`, this rather moves a bunch of things behind a `get_db_session` context-manager.
That way, it should not be possible anymore (or rather, less likely?) to leak an open `db_session`.