Clean up `requirements` and add `.envrc` (#318)
The `.envrc` file in conjunction with `direnv` will allow to automatically set up and activate a local `.venv` which allows running `pytest` and friends locally outside of docker.
This local `.venv` also fixes my local vscode editor completions and type hints, which is a huge plus.

Another change here is that this removes `django-admin`. I don’t see this being used anywhere, and it fails to compile locally, as Python 3.12 removed some deprecated functionality that is required for building `screen` locally, which is being pulled in transitively via `django-admin`.
Why this all works in docker though is a mystery to me.

Otherwise it just alphabetically sorts all the various requirements files/lists.