chore: rename bundle_analysis.bundle_analysis_threshold (#323)
There were 2 configs with the same name: `comment.bundle_analysis_threshold`
and `bundle_analysis.bundle_analysis_threshold`. They are similar but will be kept
separate. One is the threshold to create a comment, the other to emit warnings or fail
a status.

To help differentiate them we are renaming one. Because `comment.bundle_analysis_threshold` is already used it's easier to rename the other one.