chore: adding more causes for InvalidInstallationError (#333)
* chore: adding more causes for InvalidInstallationError

We are still being hit by many invalid installation errors.
The `permission_error` real cause is still not unclear. At first the suspicion
was suspended apps, but that is not the whole story (from empirical data).

I suspected rate limit errors, and we discovered an unconfigured cache at some point.
That has been remedied but the errors are still going strong.

These changes attempt to further breakdown the responses we get to help with debugging.
Mostly it will verify if the issue is indeed with rate limit or not.

* fix test