chore: mark suspended apps so (#334)
In commit 8f92394bc15885ffc02f7003cccd2cd246dbbc76 we added a further breakdown
of why `InvalidInstallationError` was raised when trying to get the access_token.
In particular the body of the response tells us if the installation is suspended,
which is quite convenient.

Scrolling through the events in Sentry for that issue (example)[https://codecov.sentry.io/issues/5619600100/events/d9d0c0f6e66a43dba92e4d0021b9f0fa/?project=4165036&query=is%3Aunresolved+issue.priority%3A%5Bhigh%2C+medium%5D&referrer=previous-event&statsPeriod=7d&stream_index=10&utc=true], there's quite a few "installation_suspended" entries.

We do have a mechanism to mark installations as suspended, so we don't try to get tokens for those, but it relies on webhooks, and these installations were probably suspended before that was put in place.

So these changes mark installations that are suspended as suspended after we get the error.

To unsuspend an installation is a manual task done by the user in the GitHub UI. It sends us a webhook and we mark the app as unsuspended.