chore: reduce prefech limit (#339)
see docs: https://docs.celeryq.dev/en/stable/userguide/optimizing.html#prefetch-limits

We want to reudce the number of tasks that are prefetched by workers, following celery docs suggestion:

> If you have many tasks with a long duration you want the multiplier value to be one: meaning it’ll only reserve one task per worker process at a time.

Most of our tasks are long running, and we have experienced a considerable number of tasks
being "in a limbo". So reducing the number of prefetched tasks might help with that too.