chore: use graphQL to get PR info (github) (#381)
Changes the endpoint to get a single PR info to use graphQL api.
The goal is to slightly improve the baseline time of the request and
reduce the difference a little from the fastest request to p90.

For more details on the exploration see https://github.com/codecov/engineering-team/issues/2627