Adjust docker images and direnv config (#430)
This simplifies the `direnv` config a bit. As `uv sync` is so fast, we can just unconditionally run it. A local `uv sync` has also removed some dependencies from the lockfile which have been removed from the dependencies.

Also adjust the docker images, downgrading `redis` to match the version we are actually running, and picking a lighter-weight `timescaledb` image which should be quicker to download.